chore(refactor): add folder for branding tests, move files around
