+++ tests_cypress/cypress/e2e/admin/all.cy.js
@@ -2,4 +2,4 @@
 import './login.cy';
 import './register.cy';
 import './qualtrics.cy';
-import './request_branding.cy';
+import './branding/request_branding.cy';
