+++ tests_cypress/cypress/e2e/admin/branding/all.cy.js
@@ -1,0 +1,2 @@
+import './a11y.cy';
+import './request_branding.cy';
